Split accessors into separate files for ease of maintenance.
This way we can start to detect some overlap between them and
better manage their relationships.
