+++ src/index.ts
@@ -2,6 +2,12 @@ export * from "./form";
 export * from "./converter";
 export * from "./converters";
 export * from "./accessor";
+export * from "./field-accessor";
+export * from "./form-accessor";
+export * from "./repeating-form-accessor";
+export * from "./repeating-form-indexed-accessor";
+export * from "./sub-form-accessor";
+export * from "./validation-props";
 export * from "./controlled";
 export * from "./antd";
 export * from "./state";
