+++ src/validation-props.ts
@@ -1,0 +1,13 @@
+import { FieldAccessor } from "./field-accessor";
+
+export interface ValidationProps {
+  (accessor: FieldAccessor<any, any, any>): object;
+}
+
+export let currentValidationProps: ValidationProps = () => {
+  return {};
+};
+
+export function setupValidationProps(validationProps: ValidationProps) {
+  currentValidationProps = validationProps;
+}
